+++ modules-template/prov-template-compiler/src/main/java/org/openprovenance/prov/template/compiler/past/checker/TypeChecker.java
@@ -256,10 +256,6 @@ public class TypeChecker {
                     : EnumSet.copyOf(def.modifiers);
             env.define(varName, def.type, mods);
         }
-
-        if ("sb".equals(varName)) {
-            System.out.println("Registered variable 'sb' of type " + def.type + " in " + className + "." + methodName + " with ENV " + env.getVariables());
-        }
     }
 
     // --- Assignment ---